requirements.txt
ultralytics
fastapi
uvicorn
python-multipart
pillow
requests
